chore: get json schema from db not frontend
